EP-1518 При проведении сверки по контингенту должны участвовать поля: код льготы, статус льготы, даты действия льготы (даты с пометкой ДТСЗН).

+++ processor/src/core/java/ru/axetta/ecafe/processor/core/persistence/utils/DAOUtils.java
@@ -3938,7 +3938,7 @@ public class DAOUtils {
 
     public static CategoryDiscountDSZN getCategoryDiscountDSZNByDSZNCode(Session session, Long dsznCode) {
         Criteria criteria = session.createCriteria(CategoryDiscountDSZN.class);
-        criteria.add(Restrictions.eq("code", dsznCode));
+        criteria.add(Restrictions.eq("code", dsznCode.intValue()));
         return (CategoryDiscountDSZN) criteria.uniqueResult();
     }
